+++ src/fbdump.c
@@ -5,6 +5,7 @@
 #include "drm_disp.h"
 #include <stdio.h>
 #include <string.h>
+#include <stdlib.h>
 #include <unistd.h>
 
 int main(int argc, char **argv)